--- test/beans-market-test.js
@@ -1,1632 +1,0 @@
-const { expect } = require("chai");
-const { ethers } = require("hardhat");
-const { loadFixture } = require("@nomicfoundation/hardhat-network-helpers");
-
-const ONE_ETH = ethers.utils.parseEther("1.0");
-const BIG_ZERO = ethers.constants.Zero;
-const ADDR_ZERO = ethers.constants.AddressZero;
-
-describe("Beanie Market", function () {
-  async function deployMarketAndNFTFixture() {
-    const [owner, ...addrs] = await ethers.getSigners();
-    const admin = addrs[9];
-
-    const ERC20 = await ethers.getContractFactory("ERC20Mock");
-    const paymentToken = await ERC20.deploy();
-    await paymentToken.deployed();
-
-    for (let i = 0; i<5; i++) {
-      paymentToken.mint(addrs[i].address, ONE_ETH.mul(100));
-    }
-
-    const MARKET = await ethers.getContractFactory("BeanieMarketV11");
-    const beanieMarket = await MARKET.deploy(paymentToken.address);
-    await beanieMarket.deployed();
-
-    const NFT = await ethers.getContractFactory("ERC721Mock");
-    const dummyNFT = await NFT.deploy();
-    await dummyNFT.deployed();
-
-    await dummyNFT.connect(addrs[0]).mint(addrs[0].address, 10);
-    await dummyNFT.connect(addrs[1]).mint(addrs[1].address, 10);
-
-    await beanieMarket.connect(owner).setCollectionTrading(dummyNFT.address, true);
-    await beanieMarket.connect(owner).setCollectionOwner(dummyNFT.address, addrs[5].address);;
-    await beanieMarket.connect(owner).setCollectionOwnerFee(dummyNFT.address, 100); //1% fee
-
-    const block = await ethers.provider.getBlock();
-    const now = block['timestamp']
-
-    return { beanieMarket, dummyNFT, paymentToken, owner, admin, addrs, now };
-  }
-
-  async function deployMarketAndListNFTsFixture() {
-    const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndNFTFixture);
-
-    await dummyNFT.connect(addrs[0]).setApprovalForAll(beanieMarket.address, true);
-    await dummyNFT.connect(addrs[1]).setApprovalForAll(beanieMarket.address, true);
-    await beanieMarket.connect(addrs[0]).listToken(dummyNFT.address, 1, ONE_ETH, now + 100);
-    await beanieMarket.connect(addrs[0]).listToken(dummyNFT.address, 2, ONE_ETH, now + 1000);
-    await beanieMarket.connect(addrs[0]).listToken(dummyNFT.address, 3, ONE_ETH, now + 1000);
-    await beanieMarket.connect(addrs[0]).listToken(dummyNFT.address, 4, ONE_ETH, now + 1000);
-    await beanieMarket.connect(addrs[1]).listToken(dummyNFT.address, 11, ONE_ETH, now + 10000);
-
-    return { beanieMarket, dummyNFT, paymentToken, owner, addrs, now };
-  }
-
-  async function deployMarketAndMakeOffersFixture() {
-    const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndNFTFixture);
-    await paymentToken.connect(addrs[2]).approve(beanieMarket.address, ONE_ETH.mul(10));
-    await paymentToken.connect(addrs[3]).approve(beanieMarket.address, ONE_ETH.mul(10));
-    await beanieMarket.connect(addrs[2]).makeOffer(dummyNFT.address, 1, ONE_ETH, now + 100);
-    await beanieMarket.connect(addrs[2]).makeOffer(dummyNFT.address, 1, ONE_ETH, now + 100);
-    await beanieMarket.connect(addrs[2]).makeOffer(dummyNFT.address, 2, ONE_ETH, now + 1000);
-    await beanieMarket.connect(addrs[3]).makeOffer(dummyNFT.address, 1, ONE_ETH, now + 10000);
-    await beanieMarket.connect(addrs[3]).makeOffer(dummyNFT.address, 3, ONE_ETH, now + 10000);
-
-    return { beanieMarket, dummyNFT, paymentToken, owner, addrs, now };
-  }
-
-  async function deployMarketAndMakeEscrowOffersFixture() {
-    const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndNFTFixture);
-
-    await beanieMarket.connect(addrs[2]).makeEscrowedOffer(
-      dummyNFT.address,
-      1,
-      now + 100,
-      {value: ONE_ETH}
-    );
-    await beanieMarket.connect(addrs[2]).makeEscrowedOffer(
-      dummyNFT.address,
-      1,
-      now + 100,
-      {value: ONE_ETH.mul(2)}
-    );
-    await beanieMarket.connect(addrs[2]).makeEscrowedOffer(
-      dummyNFT.address,
-      2,
-      now + 1000,
-      {value: ONE_ETH.mul(3)}
-    );
-    await beanieMarket.connect(addrs[3]).makeEscrowedOffer(
-      dummyNFT.address,
-      1,
-      now + 10000,
-      {value: ONE_ETH}
-    );
-    await beanieMarket.connect(addrs[3]).makeEscrowedOffer(
-      dummyNFT.address,
-      3,
-      now + 10000,
-      {value: ONE_ETH}
-    );
-
-    return { beanieMarket, dummyNFT, paymentToken, owner, addrs, now };
-  }
-
-  describe("Deployment", function () {
-    it("Deployment", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs } = await loadFixture(deployMarketAndNFTFixture);
-      expect(await beanieMarket.TOKEN()).to.equal(paymentToken.address);
-      expect(await dummyNFT.symbol()).to.equal("DUMMY")
-      expect(await dummyNFT.balanceOf(addrs[0].address)).to.eql(ethers.BigNumber.from(10));
-    });
-  })
-
-  describe("Listings", function () {
-    it("List token and updated storage structures", async function () {
-      const { beanieMarket, dummyNFT, owner, addrs, now } = await loadFixture(deployMarketAndNFTFixture);
-      const address0 = addrs[0];
-      const address1 = addrs[1];
-
-      await dummyNFT.connect(address0).setApprovalForAll(beanieMarket.address, true);
-      await dummyNFT.connect(address1).setApprovalForAll(beanieMarket.address, true);
-
-      //Check the storage structures of several listings. We do this for two different lister addresses to test.
-      //TODO: maybe add a second contract address.
-      await beanieMarket.connect(address0).listToken(dummyNFT.address, 1, ONE_ETH, now + 10);
-      let listingsByLister = await beanieMarket.getListingsByLister(address0.address);
-      let listingsByContract = await beanieMarket.getListingsByContract(dummyNFT.address);
-      expect(listingsByLister.length).to.equal(1)
-      expect(listingsByContract.length).to.equal(1)
-      expect(listingsByLister[0]).to.equal(listingsByContract[0]);
-      expect(await beanieMarket.listings(listingsByLister[0])).to.eql(
-        [
-          ethers.BigNumber.from(1),
-          ONE_ETH,
-          ethers.BigNumber.from(now+10),
-          dummyNFT.address,
-          addrs[0].address
-        ]
-      )
-
-      await beanieMarket.connect(address0).listToken(dummyNFT.address, 2, ONE_ETH, now + 100);
-      listingsByLister = await beanieMarket.getListingsByLister(address0.address);
-      listingsByContract = await beanieMarket.getListingsByContract(dummyNFT.address);
-      expect(listingsByLister.length).to.equal(2)
-      expect(listingsByContract.length).to.equal(2)
-      expect(listingsByLister[1]).to.equal(listingsByContract[1]);
-      expect(await beanieMarket.listings(listingsByLister[1])).to.eql(
-        [
-          ethers.BigNumber.from(2),
-          ONE_ETH,
-          ethers.BigNumber.from(now+100),
-          dummyNFT.address,
-          addrs[0].address
-        ]
-      )
-
-      await beanieMarket.connect(address1).listToken(dummyNFT.address, 11, ONE_ETH, now + 1000);
-      listingsByLister = await beanieMarket.getListingsByLister(address1.address);
-      listingsByContract = await beanieMarket.getListingsByContract(dummyNFT.address);
-      expect(listingsByLister[0]).to.equal(listingsByContract[2]);
-      expect(listingsByLister.length).to.equal(1)
-      expect(listingsByContract.length).to.equal(3)
-      expect(await beanieMarket.listings(listingsByLister[0])).to.eql(
-        [
-          ethers.BigNumber.from(11),
-          ONE_ETH,
-          ethers.BigNumber.from(now+1000),
-          dummyNFT.address,
-          addrs[1].address
-        ]
-      )
-
-    });
-
-    it("Fulfill listing ownership change", async function () {
-      const { beanieMarket, dummyNFT, owner, addrs, now } = await loadFixture(deployMarketAndListNFTsFixture);
-      const address0 = addrs[0];
-
-      expect(await dummyNFT.ownerOf(1)).to.equal(addrs[0].address);
-      const listingIds = await beanieMarket.getListingsByContract(dummyNFT.address);
-      const listingToFulfill = listingIds[0];
-      await beanieMarket.connect(addrs[1]).fulfillListing(listingToFulfill, addrs[1].address, {value: ONE_ETH});
-
-      expect(await dummyNFT.ownerOf(1)).to.equal(addrs[1].address);
-    });
-
-    //When listing0 is fulfilled, new posInListingByLister[0] should be previous last listing
-
-    it("Fulfill listing updates storage structures", async function () {
-      const { beanieMarket, dummyNFT, owner, addrs, now } = await loadFixture(deployMarketAndListNFTsFixture);
-      const address0 = addrs[0];
-
-      expect(await dummyNFT.ownerOf(1)).to.equal(addrs[0].address);
-      const listingIds = await beanieMarket.getListingsByContract(dummyNFT.address);
-      const listingToFulfill = listingIds[0];
-
-      let listingsByLister0 = await beanieMarket.getListingsByLister(addrs[0].address);
-      let listingsByLister1 = await beanieMarket.getListingsByLister(addrs[1].address);
-      let listingsByContract = await beanieMarket.getListingsByContract(dummyNFT.address);
-      expect(listingsByLister0).to.contain(listingToFulfill)
-      expect(listingsByContract).to.contain(listingToFulfill)
-      expect(listingsByLister1).to.not.contain(listingToFulfill)
-
-      const listingTailLister = await listingsByLister0[listingsByLister0.length - 1];
-      const listingTailContract = await listingsByContract[listingsByContract.length - 1];
-
-      //Fulfill listing
-      await beanieMarket.connect(addrs[1]).fulfillListing(listingToFulfill, addrs[1].address, {value: ONE_ETH});
-      expect(await dummyNFT.ownerOf(1)).to.equal(addrs[1].address);
-
-      listingsByLister0 = await beanieMarket.getListingsByLister(addrs[0].address);
-      listingsByLister1 = await beanieMarket.getListingsByLister(addrs[1].address);
-      listingsByContract = await beanieMarket.getListingsByContract(dummyNFT.address);
-      expect(listingsByLister0.length).to.equal(3)
-      expect(listingsByLister1.length).to.equal(1)
-      expect(listingsByContract.length).to.equal(4)
-
-      //Make sure listing ID is removed from listingsByLister
-      expect(listingsByLister0).to.not.contain(listingToFulfill)
-      expect(listingsByLister1).to.not.contain(listingToFulfill)
-      expect(listingsByContract).to.not.contain(listingToFulfill)
-
-      //Check to see if new listingsByLister and listingsByContract is correct
-      expect((await beanieMarket.posInListings(listingTailLister)).posInListingsByLister).to.eql(BIG_ZERO);
-      expect((await beanieMarket.posInListings(listingTailContract)).posInListingsByContract).to.eql(BIG_ZERO);
-
-      //Make sure old listing entry has been removed
-      expect(await beanieMarket.listings(listingToFulfill)).to.eql([BIG_ZERO, BIG_ZERO, BIG_ZERO, ADDR_ZERO, ADDR_ZERO]);
-      expect(await beanieMarket.posInListings(listingToFulfill)).to.eql([BIG_ZERO, BIG_ZERO]);
-    });
-
-    it("Fulfill listing updates storage structures (single-length listing array)", async function () {
-      const { beanieMarket, dummyNFT, owner, addrs, now } = await loadFixture(deployMarketAndListNFTsFixture);
-      const address0 = addrs[0];
-
-      expect(await dummyNFT.ownerOf(11)).to.equal(addrs[1].address);
-      const listingIds = await beanieMarket.getListingsByLister(addrs[1].address);
-      const listingToFulfill = listingIds[0];
-
-      await beanieMarket.connect(addrs[1]).fulfillListing(listingToFulfill, addrs[2].address, {value: ONE_ETH});
-      expect(await dummyNFT.ownerOf(11)).to.equal(addrs[2].address);
-      expect(await beanieMarket.listings(listingToFulfill)).to.eql([BIG_ZERO, BIG_ZERO, BIG_ZERO, ADDR_ZERO, ADDR_ZERO]);
-      expect(await beanieMarket.posInListings(listingToFulfill)).to.eql([BIG_ZERO, BIG_ZERO]);
-    });
-
-    it("Fulfill listing feesOff sends correct eth amount", async function () {
-      const { beanieMarket, dummyNFT, owner, addrs, now } = await loadFixture(deployMarketAndListNFTsFixture);
-      const address0 = addrs[0];
-
-      const listingIds = await beanieMarket.getListingsByContract(dummyNFT.address);
-      const listingToFulfill = listingIds[0];
-
-      await beanieMarket.connect(owner).setFeesOn(false);
-
-      let oldOwnerBalanceBefore = await addrs[0].getBalance();
-      let newOwnerBalanceBefore = await addrs[1].getBalance();
-
-      let tx = await beanieMarket.connect(addrs[1]).fulfillListing(listingToFulfill, addrs[1].address, {value: ONE_ETH});
-      let receipt = await tx.wait();
-      const gasSpent = receipt.gasUsed.mul(receipt.effectiveGasPrice);
-
-      let oldOwnerBalanceAfter = await addrs[0].getBalance();
-      let newOwnerBalanceAfter = await addrs[1].getBalance();
-
-      expect(oldOwnerBalanceAfter).to.eql(oldOwnerBalanceBefore.add(ONE_ETH));
-      expect(newOwnerBalanceAfter).to.eql(newOwnerBalanceBefore.sub(ONE_ETH).sub(gasSpent));
-
-      expect(await dummyNFT.ownerOf(1)).to.equal(addrs[1].address);
-
-    });
-
-    it("Fulfill listing feesOn sends correct token amount, autosend on", async function () {
-      const { beanieMarket, dummyNFT, owner, addrs, now } = await loadFixture(deployMarketAndListNFTsFixture);
-      const address0 = addrs[0];
-      await beanieMarket.connect(owner).setAutoSendFees(true);
-
-      const listingIds = await beanieMarket.getListingsByContract(dummyNFT.address);
-      const listingToFulfill = listingIds[0];
-
-      const devFee = await beanieMarket.devFee();
-      const beanieHolderFee = await beanieMarket.beanieHolderFee();
-      const beanBuybackFee = await beanieMarket.beanBuybackFee();
-      const collectionOwnerFee = await beanieMarket.getCollectionFee(dummyNFT.address);
-      const totalFee = devFee.add(beanieHolderFee).add(beanBuybackFee).add(collectionOwnerFee);
-
-      const devAddress = await beanieMarket.devAddress();
-      const beanHolderAddress = await beanieMarket.beanieHolderAddress();
-      const beanBuybackAddress = await beanieMarket.beanBuybackAddress();
-
-      let oldOwnerBalanceBefore = await addrs[0].getBalance();
-      let newOwnerBalanceBefore = await addrs[1].getBalance();
-      let collectionOwnerBalBefore = await addrs[5].getBalance();
-
-      let tx = await beanieMarket.connect(addrs[1]).fulfillListing(listingToFulfill, addrs[1].address, {value: ONE_ETH});
-      let receipt = await tx.wait();
-      const gasSpent = receipt.gasUsed.mul(receipt.effectiveGasPrice);
-
-      let oldOwnerBalanceAfter = await addrs[0].getBalance();
-      let newOwnerBalanceAfter = await addrs[1].getBalance();
-      let collectionOwnerBalAfter = await addrs[5].getBalance();
-
-      let devFeeAmount = ONE_ETH.mul(devFee).div(10000);
-      let beanieHolderFeeAmount = ONE_ETH.mul(beanieHolderFee).div(10000);
-      let beanBuybackFeeAmount = ONE_ETH.mul(beanBuybackFee).div(10000);
-      let collectionOwnerFeeAmount = ONE_ETH.mul(collectionOwnerFee).div(10000);
-      let afterFeePrice = ONE_ETH.mul(totalFee).div(10000);
-
-      //Check principal balances
-      expect(oldOwnerBalanceAfter).to.eql(oldOwnerBalanceBefore.add(ONE_ETH).sub(afterFeePrice));
-      expect(newOwnerBalanceAfter).to.eql(newOwnerBalanceBefore.sub(ONE_ETH).sub(gasSpent));
-
-      //Check fee balances
-      expect(await ethers.provider.getBalance(devAddress)).to.eql(devFeeAmount);
-      expect(await ethers.provider.getBalance(beanHolderAddress)).to.eql(beanieHolderFeeAmount);
-      expect(await ethers.provider.getBalance(beanBuybackAddress)).to.eql(beanBuybackFeeAmount);
-      expect(await collectionOwnerBalAfter).to.eql(collectionOwnerBalBefore.add(collectionOwnerFeeAmount));
-
-      expect(await dummyNFT.ownerOf(1)).to.equal(addrs[1].address);
-
-    });
-
-    it("Fulfill listing feesOn sends correct token amount, autosend off and then process", async function () {
-      const { beanieMarket, dummyNFT, owner, addrs, now } = await loadFixture(deployMarketAndListNFTsFixture);
-      const address0 = addrs[0];
-      await beanieMarket.connect(owner).setAutoSendFees(false);
-
-      const listingIds = await beanieMarket.getListingsByContract(dummyNFT.address);
-      const listingToFulfill = listingIds[0];
-
-      //Get fee percentages
-      const devFee = await beanieMarket.devFee();
-      const beanieHolderFee = await beanieMarket.beanieHolderFee();
-      const beanBuybackFee = await beanieMarket.beanBuybackFee();
-      const collectionOwnerFee = await beanieMarket.getCollectionFee(dummyNFT.address);
-      const accruedAdminFees = devFee.add(beanieHolderFee).add(beanBuybackFee);
-      const totalFee = accruedAdminFees.add(collectionOwnerFee);
-
-      const devAddress = await beanieMarket.devAddress();
-      const beanHolderAddress = await beanieMarket.beanieHolderAddress();
-      const beanBuybackAddress = await beanieMarket.beanBuybackAddress();
-
-      let oldOwnerBalanceBefore = await addrs[0].getBalance();
-      let newOwnerBalanceBefore = await addrs[1].getBalance();
-      let collectionOwnerBalBefore = await addrs[5].getBalance();
-
-      let tx = await beanieMarket.connect(addrs[1]).fulfillListing(listingToFulfill, addrs[1].address, {value: ONE_ETH});
-      let receipt = await tx.wait();
-      const gasSpent = receipt.gasUsed.mul(receipt.effectiveGasPrice);
-
-      expect(await beanieMarket.accruedAdminFeesEth()).to.eql(ONE_ETH.mul(accruedAdminFees).div(10000));
-
-      let oldOwnerBalanceAfter = await addrs[0].getBalance();
-      let newOwnerBalanceAfter = await addrs[1].getBalance();
-      let collectionOwnerBalAfter = await addrs[5].getBalance();
-
-      let collectionOwnerFeeAmount = ONE_ETH.mul(collectionOwnerFee).div(10000);
-
-      //Ensure collection owners always get auto-forwarded
-      expect(await collectionOwnerBalAfter).to.eql(collectionOwnerBalBefore.add(collectionOwnerFeeAmount));
-
-      //Check the rest of the fees
-      expect(await ethers.provider.getBalance(devAddress)).to.eql(BIG_ZERO);
-      expect(await ethers.provider.getBalance(beanHolderAddress)).to.eql(BIG_ZERO);
-      expect(await ethers.provider.getBalance(beanBuybackAddress)).to.eql(BIG_ZERO);
-
-      await beanieMarket.connect(owner).processDevFeesEth();
-
-      let devFeeAmount = ONE_ETH.mul(devFee).div(10000);
-      let beanieHolderFeeAmount = ONE_ETH.mul(beanieHolderFee).div(10000);
-      let beanBuybackFeeAmount = ONE_ETH.mul(beanBuybackFee).div(10000)
-      let afterFeePrice = ONE_ETH.mul(totalFee).div(10000);
-
-      //Check principal balances
-      expect(oldOwnerBalanceAfter).to.eql(oldOwnerBalanceBefore.add(ONE_ETH).sub(afterFeePrice));
-      expect(newOwnerBalanceAfter).to.eql(newOwnerBalanceBefore.sub(ONE_ETH).sub(gasSpent));
-
-      //Check fee balances
-      expect(await ethers.provider.getBalance(devAddress)).to.eql(devFeeAmount);
-      expect(await ethers.provider.getBalance(beanHolderAddress)).to.eql(beanieHolderFeeAmount);
-      expect(await ethers.provider.getBalance(beanBuybackAddress)).to.eql(beanBuybackFeeAmount);
-
-      expect(await dummyNFT.ownerOf(1)).to.equal(addrs[1].address);
-
-    });
-
-    it("Delist token", async function () {
-      const { beanieMarket, dummyNFT, owner, addrs } = await loadFixture(deployMarketAndListNFTsFixture);
-      const address0 = addrs[0];
-      const listingIds = await beanieMarket.getListingsByContract(dummyNFT.address);
-      const listingToDelist = listingIds[0];
-
-      let listingsByLister0 = await beanieMarket.getListingsByLister(addrs[0].address);
-      let listingsByLister1 = await beanieMarket.getListingsByLister(addrs[1].address);
-      let listingsByContract = await beanieMarket.getListingsByContract(dummyNFT.address);
-      expect(listingsByLister0.length).to.equal(4)
-      expect(listingsByLister1.length).to.equal(1)
-      expect(listingsByContract.length).to.equal(5)
-
-      await beanieMarket.connect(address0).delistToken(listingToDelist);
-
-      listingsByLister0 = await beanieMarket.getListingsByLister(addrs[0].address);
-      listingsByLister1 = await beanieMarket.getListingsByLister(addrs[1].address);
-      listingsByContract = await beanieMarket.getListingsByContract(dummyNFT.address);
-      expect(listingsByLister0).to.not.contain(listingToDelist)
-      expect(listingsByLister1).to.not.contain(listingToDelist)
-      expect(listingsByContract).to.not.contain(listingToDelist)
-
-      expect(listingsByLister0.length).to.equal(3)
-      expect(listingsByLister1.length).to.equal(1)
-      expect(listingsByContract.length).to.equal(4)
-
-      //Check to see if new listingsByLister and listingsByContract is correct
-      expect(await beanieMarket.listings(listingToDelist)).to.eql([BIG_ZERO, BIG_ZERO, BIG_ZERO, ADDR_ZERO, ADDR_ZERO]);
-      expect(await beanieMarket.posInListings(listingToDelist)).to.eql([BIG_ZERO, BIG_ZERO]);
-    });
-
-    it("Creating a new listing for same token removes old listing.", async function () {
-      const { beanieMarket, dummyNFT, owner, addrs, now } = await loadFixture(deployMarketAndListNFTsFixture);
-      const listingIds = await beanieMarket.getListingsByContract(dummyNFT.address);
-      const listingHash = listingIds[0];
-      const listingData = await beanieMarket.listings(listingHash);
-      const currentListingOrderHash = await beanieMarket.currentListingOrderHash(listingData.contractAddress, listingData.tokenId);
-      expect(currentListingOrderHash).to.equal(listingHash);
-
-      await beanieMarket.connect(addrs[0]).listToken(listingData.contractAddress, listingData.tokenId, ONE_ETH, now + 100);
-
-      const listingIdsNew = await beanieMarket.getListingsByContract(dummyNFT.address);
-      const currentListingOrderHashNew = await beanieMarket.currentListingOrderHash(listingData.contractAddress, listingData.tokenId);
-      const listingDataNew = await beanieMarket.listings(currentListingOrderHashNew);
-
-      //Old listing has been canceled
-      expect(await beanieMarket.listings(listingHash)).to.eql([BIG_ZERO, BIG_ZERO, BIG_ZERO, ADDR_ZERO, ADDR_ZERO])
-      expect(await beanieMarket.posInListings(listingHash)).to.eql([BIG_ZERO, BIG_ZERO]);
-
-      expect(listingData.contractAddress, listingData.tokenId).to.equal(listingDataNew.contractAddress, listingDataNew.tokenId)
-      expect(listingIdsNew).to.not.contain(currentListingOrderHash);
-      expect(listingIdsNew).to.contain(currentListingOrderHashNew);
-      expect(currentListingOrderHash).to.not.equal(currentListingOrderHashNew);
-    });
-  });
-
-  describe("Listing token errors", function () {
-    it("Cannot list an unowned token", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndNFTFixture);
-      await expect(beanieMarket.connect(addrs[5]).listToken(dummyNFT.address, 4, ONE_ETH, now + 1000)
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANCallerNotOwner");
-    });
-    
-    it("Cannot list token if contract is not approved", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndNFTFixture);
-      await expect(beanieMarket.connect(addrs[0]).listToken(dummyNFT.address, 4, ONE_ETH, now + 1000)
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANContractNotApproved");
-    });
-
-    it("Cannot list token if expiry is before current timestamp", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs } = await loadFixture(deployMarketAndNFTFixture);
-      const blockNumBefore = await ethers.provider.getBlockNumber();
-      const blockBefore = await ethers.provider.getBlock(blockNumBefore);
-      const now = blockBefore.timestamp;
-
-      await dummyNFT.connect(addrs[0]).setApprovalForAll(beanieMarket.address, true);
-
-      await expect(beanieMarket.connect(addrs[0]).listToken(dummyNFT.address, 4, ONE_ETH, now-1)
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANBadExpiry");
-      await expect(beanieMarket.connect(addrs[0]).listToken(dummyNFT.address, 4, ONE_ETH, now)
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANBadExpiry");
-      await beanieMarket.connect(addrs[0]).listToken(dummyNFT.address, 4, ONE_ETH, now+100);
-    });
-  })
-
-  describe.skip("Fulfill listing token errors", function () {
-    it("Cannot fulfill listing for an unlisted token", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndNFTFixture);
-      await expect(beanieMarket.connect(addrs[5]).fulfillListing(ethers.utils.hexZeroPad(0x2, 32), addrs[5].address)
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANCollectionNotEnabled");
-    });
-
-    it("Cannot fulfill listing by sending too little ETH", async function () {
-      const { beanieMarket, dummyNFT, owner, addrs, now } = await loadFixture(deployMarketAndListNFTsFixture);
-      const address0 = addrs[0];
-
-      expect(await dummyNFT.ownerOf(1)).to.equal(addrs[0].address);
-      const listingIds = await beanieMarket.getListingsByContract(dummyNFT.address);
-      const listingToFulfill = listingIds[0];
-      const listingData = await beanieMarket.listings(listingToFulfill);
-
-      await expect(beanieMarket.connect(addrs[5]).fulfillListing(
-        listingToFulfill, addrs[1].address, {value: listingData.price.div(2)})
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANNotEnoughEthSent");
-    });
-
-    it("Cannot fulfill listing if past expiry", async function () {
-      const { beanieMarket, dummyNFT, owner, addrs, now } = await loadFixture(deployMarketAndListNFTsFixture);
-      const address0 = addrs[0];
-
-      expect(await dummyNFT.ownerOf(1)).to.equal(addrs[0].address);
-      const listingIds = await beanieMarket.getListingsByContract(dummyNFT.address);
-      const listingToFulfill = listingIds[0];
-      const listingData = await beanieMarket.listings(listingToFulfill);
-
-      await network.provider.send("evm_setNextBlockTimestamp", [Number(listingData.expiry) + 10])
-      await network.provider.send("evm_mine")
-
-      await expect(beanieMarket.connect(addrs[5]).fulfillListing(
-        listingToFulfill, addrs[1].address, {value: listingData.price})
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANOrderExpired");
-    });
-
-    it("Cannot fulfill listing if token owner has changed", async function () {
-      const { beanieMarket, dummyNFT, owner, addrs, now } = await loadFixture(deployMarketAndListNFTsFixture);
-      const address0 = addrs[0];
-
-      expect(await dummyNFT.ownerOf(1)).to.equal(addrs[0].address);
-
-      const listingIds = await beanieMarket.getListingsByContract(dummyNFT.address);
-      const listingToFulfill = listingIds[0];
-      const listingData = await beanieMarket.listings(listingToFulfill);
-
-      await dummyNFT.connect(addrs[0]).transferFrom(addrs[0].address, addrs[1].address, listingData.tokenId);
-
-      await expect(beanieMarket.connect(addrs[5]).fulfillListing(
-        listingToFulfill, addrs[4].address, {value: listingData.price})
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANListingNotActive");
-    });
-  })
-
-  describe.skip("Delist token errors and delist cases", function () { // works, fixtures just break -- hh problem
-    it("Cannot delist unowned token", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndListNFTsFixture);
-
-      const listings = await beanieMarket.getListingsByContract(dummyNFT.address);
-      const listing0hash = listings[0];
-      const listing0Data = await beanieMarket.listings(listing0hash);
-
-      await expect(beanieMarket.connect(addrs[5]).delistToken(listing0hash)
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANDelistNotApproved");
-    });
-
-    it("Can delist unowned token if caller is admin", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndListNFTsFixture);
-
-      const listings = await beanieMarket.getListingsByContract(dummyNFT.address);
-      const listing0hash = listings[0];
-      const listing0Data = await beanieMarket.listings(listing0hash);
-
-      await beanieMarket.connect(owner).delistToken(listing0hash);
-    });
-
-    it("Can delist unowned token if lister is no longer token owner.", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndListNFTsFixture);
-
-      const listings = await beanieMarket.getListingsByContract(dummyNFT.address);
-      const listing0hash = listings[0];
-      const listing0Data = await beanieMarket.listings(listing0hash);
-
-      await expect(beanieMarket.connect(addrs[2]).delistToken(listing0hash)
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANDelistNotApproved");
-
-      await dummyNFT.connect(addrs[0]).transferFrom(addrs[0].address, addrs[1].address, listing0Data.tokenId);
-      await beanieMarket.connect(addrs[2]).delistToken(listing0hash);
-    });
-
-    it("Can delist unowned token if expiry has passed.", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndListNFTsFixture);
-
-      const listings = await beanieMarket.getListingsByContract(dummyNFT.address);
-      const listing0hash = listings[0];
-      const listing0Data = await beanieMarket.listings(listing0hash);
-
-      await expect(beanieMarket.connect(addrs[2]).delistToken(listing0hash)
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANDelistNotApproved");
-
-      await network.provider.send("evm_setNextBlockTimestamp", [Number(listing0Data.expiry) - 10])
-      await network.provider.send("evm_mine")
-
-      await expect(beanieMarket.connect(addrs[2]).delistToken(listing0hash)
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANDelistNotApproved");
-
-      await network.provider.send("evm_setNextBlockTimestamp", [Number(listing0Data.expiry) + 1])
-      await network.provider.send("evm_mine")
-
-      await beanieMarket.connect(addrs[2]).delistToken(listing0hash);
-    });
-  })
-
-  describe("non-escrow offers", function () {
-    it("Make non-escrow offer", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndNFTFixture);
-
-      //Check the storage structures of several listings. We do this for two different lister addresses to test.
-      //TODO: maybe add a second contract address.
-      await paymentToken.connect(addrs[2]).approve(beanieMarket.address, ONE_ETH.mul(10));
-      await beanieMarket.connect(addrs[2]).makeOffer(dummyNFT.address, 1, ONE_ETH, now + 10); // offer1
-
-      const addr2Offers_1 = await beanieMarket.getOffersByOfferer(addrs[2].address)
-      expect(addr2Offers_1.length).to.equal(1);
-
-      //same params, make sure it works
-      await beanieMarket.connect(addrs[2]).makeOffer(dummyNFT.address, 1, ONE_ETH, now + 10);
-      const addr2Offers_2 = await beanieMarket.getOffersByOfferer(addrs[2].address)
-      expect(addr2Offers_2.length).to.equal(2);
-      expect(addr2Offers_2[0]).to.not.equal(addr2Offers_2[1]);
-
-      //Offer for a different token
-      await beanieMarket.connect(addrs[2]).makeOffer(dummyNFT.address, 2, ONE_ETH, now + 100); //offer2
-      const addr2Offers_3 = await beanieMarket.getOffersByOfferer(addrs[2].address)
-      expect(addr2Offers_3.length).to.equal(3);
-
-      //3rd account makes an offer
-      await paymentToken.connect(addrs[3]).approve(beanieMarket.address, ONE_ETH.mul(10));
-      await beanieMarket.connect(addrs[3]).makeOffer(dummyNFT.address, 1, ONE_ETH, now + 1000); //offer3
-      const addr3Offers_1 = await beanieMarket.getOffersByOfferer(addrs[3].address);
-      expect(addr3Offers_1.length).to.equal(1);
-
-      const offer1 = await beanieMarket.offers(addr2Offers_3[0]);
-      const offer2 = await beanieMarket.offers(addr2Offers_3[2]);
-      const offer3 = await beanieMarket.offers(addr3Offers_1[0]);
-
-      expect(offer1).to.eql(
-        [
-          ethers.BigNumber.from(1),
-          ONE_ETH,
-          ethers.BigNumber.from(now + 10),
-          dummyNFT.address,
-          addrs[2].address,
-          false
-        ]
-      )
-
-      expect(offer2).to.eql(
-        [
-          ethers.BigNumber.from(2),
-          ONE_ETH,
-          ethers.BigNumber.from(now + 100),
-          dummyNFT.address,
-          addrs[2].address,
-          false
-        ]
-      )
-
-      expect(offer3).to.eql(
-        [
-          ethers.BigNumber.from(1),
-          ONE_ETH,
-          ethers.BigNumber.from(now + 1000),
-          dummyNFT.address,
-          addrs[3].address,
-          false
-        ]
-      )
-    });
-
-    it("Fulfill non-escrow offer ownership change", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeOffersFixture);
-
-      const addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      const addr3offers = await beanieMarket.getOffersByOfferer(addrs[3].address);
-
-      const offer0Hash = addr2offers[0];
-      const offer0Data = await beanieMarket.offers(addr2offers[0]);
-      const offer0OldOwner = await dummyNFT.ownerOf(offer0Data.tokenId);
-
-      await dummyNFT.connect(addrs[0]).setApprovalForAll(beanieMarket.address, true);
-      await beanieMarket.connect(addrs[0]).acceptOffer(offer0Hash);
-      expect(await dummyNFT.ownerOf(offer0Data.tokenId)).to.equal(offer0Data.offerer);
-
-      const offer1Hash = addr3offers[1];
-      const offer1Data = await beanieMarket.offers(addr3offers[1]);
-      const offer1OldOwner = await dummyNFT.ownerOf(offer1Data.tokenId);
-
-      await beanieMarket.connect(addrs[0]).acceptOffer(offer1Hash);
-      expect(await dummyNFT.ownerOf(offer1Data.tokenId)).to.equal(offer1Data.offerer);
-    });
-
-    it("Fulfill non-escrow offer update storage structures", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeOffersFixture);
-
-      let addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      let addr3offers = await beanieMarket.getOffersByOfferer(addrs[3].address);
-
-      const addr2offersTail = await addr2offers[addr2offers.length - 1];
-      const addr3offersTail = await addr3offers[addr3offers.length - 1];
-
-      const offer0Hash = addr2offers[0];
-      const offer0Data = await beanieMarket.offers(addr2offers[0]);
-      const offer1Hash = addr3offers[1];
-      const offer1Data = await beanieMarket.offers(addr3offers[1]);
-
-      await dummyNFT.connect(addrs[0]).setApprovalForAll(beanieMarket.address, true);
-      await beanieMarket.connect(addrs[0]).acceptOffer(offer0Hash);
-      await beanieMarket.connect(addrs[0]).acceptOffer(offer1Hash);
-
-      addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      addr3offers = await beanieMarket.getOffersByOfferer(addrs[3].address);
-      expect(addr2offers).to.not.contain(offer0Hash)
-      expect(addr3offers).to.not.contain(offer1Hash)
-
-      expect(await beanieMarket.posInOffers(addr2offersTail)).to.eql(BIG_ZERO);
-
-    });
-
-    it("Fulfill non-escrow offer update storage structures (single length array)", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeOffersFixture);
-
-      let addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      let addr3offers = await beanieMarket.getOffersByOfferer(addrs[3].address);
-
-      const addr2offersTail = await addr2offers[addr2offers.length - 1];
-      const addr3offersTail = await addr3offers[addr3offers.length - 1];
-
-      const offer0Hash = addr2offers[0];
-      const offer0Data = await beanieMarket.offers(addr2offers[0]);
-
-      const offer1Hash = addr3offers[1];
-      const offer1Data = await beanieMarket.offers(addr3offers[1]);
-
-      await dummyNFT.connect(addrs[0]).setApprovalForAll(beanieMarket.address, true);
-      await beanieMarket.connect(addrs[0]).acceptOffer(offer0Hash);
-      await beanieMarket.connect(addrs[0]).acceptOffer(offer1Hash);
-
-      addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      addr3offers = await beanieMarket.getOffersByOfferer(addrs[3].address);
-      expect(addr2offers).to.not.contain(offer0Hash)
-      expect(addr3offers).to.not.contain(offer1Hash)
-
-      expect(await beanieMarket.posInOffers(addr2offersTail)).to.eql(BIG_ZERO);
-
-    });
-
-    it("Fulfill non-escrow offer feesOff sends correct eth amount", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeOffersFixture);
-
-      await beanieMarket.connect(owner).setFeesOn(false);
-
-      let addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      let addr3offers = await beanieMarket.getOffersByOfferer(addrs[3].address);
-
-      const offer0Hash = addr2offers[0];
-      const offer0Data = await beanieMarket.offers(addr2offers[0]);
-
-      const offer1Hash = addr3offers[1];
-      const offer1Data = await beanieMarket.offers(addr3offers[1]);
-
-      await dummyNFT.connect(addrs[0]).setApprovalForAll(beanieMarket.address, true);
-
-      let oldOwnerBalanceBefore = await paymentToken.balanceOf(addrs[0].address);
-      let newOwnerBalanceBefore = await paymentToken.balanceOf(addrs[2].address);
-
-      await beanieMarket.connect(addrs[0]).acceptOffer(offer0Hash);
-
-      let oldOwnerBalanceAfter = await paymentToken.balanceOf(addrs[0].address);
-      let newOwnerBalanceAfter = await paymentToken.balanceOf(addrs[2].address);
-
-      expect(oldOwnerBalanceAfter).to.eql(oldOwnerBalanceBefore.add(ONE_ETH));
-      expect(newOwnerBalanceAfter).to.eql(newOwnerBalanceBefore.sub(ONE_ETH));
-
-      oldOwnerBalanceBefore = await paymentToken.balanceOf(addrs[0].address);
-      newOwnerBalanceBefore = await paymentToken.balanceOf(addrs[3].address);
-
-      await beanieMarket.connect(addrs[0]).acceptOffer(offer1Hash);
-
-      oldOwnerBalanceAfter = await paymentToken.balanceOf(addrs[0].address);
-      newOwnerBalanceAfter = await paymentToken.balanceOf(addrs[3].address);
-
-      expect(oldOwnerBalanceAfter).to.eql(oldOwnerBalanceBefore.add(ONE_ETH));
-      expect(newOwnerBalanceAfter).to.eql(newOwnerBalanceBefore.sub(ONE_ETH));
-    });
-
-    it("Fulfill non-escrow feesOn sends correct token amount, autosend on", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeOffersFixture);
-
-      await beanieMarket.connect(owner).setFeesOn(true);
-      await beanieMarket.connect(owner).setAutoSendFees(true);
-
-      const devFee = await beanieMarket.devFee();
-      const beanieHolderFee = await beanieMarket.beanieHolderFee();
-      const beanBuybackFee = await beanieMarket.beanBuybackFee();
-      const collectionOwnerFee = await beanieMarket.getCollectionFee(dummyNFT.address);
-      const totalFee = devFee.add(beanieHolderFee).add(beanBuybackFee).add(collectionOwnerFee);
-
-      const devAddress = await beanieMarket.devAddress();
-      const beanHolderAddress = await beanieMarket.beanieHolderAddress();
-      const beanBuybackAddress = await beanieMarket.beanBuybackAddress();
-      const collectionOwnerAddress = await beanieMarket.getCollectionOwner(dummyNFT.address);
-
-      let addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      let addr3offers = await beanieMarket.getOffersByOfferer(addrs[3].address);
-
-      const offer0Hash = addr2offers[0];
-      const offer0Data = await beanieMarket.offers(addr2offers[0]);
-
-      const offer1Hash = addr3offers[1];
-      const offer1Data = await beanieMarket.offers(addr3offers[1]);
-
-      await dummyNFT.connect(addrs[0]).setApprovalForAll(beanieMarket.address, true);
-
-      let oldOwnerBalanceBefore = await paymentToken.balanceOf(addrs[0].address);
-      let newOwnerBalanceBefore = await paymentToken.balanceOf(addrs[2].address);
-      let collectionOwnerBalBefore = await paymentToken.balanceOf(collectionOwnerAddress);
-
-      await beanieMarket.connect(addrs[0]).acceptOffer(offer0Hash);
-
-      let oldOwnerBalanceAfter = await paymentToken.balanceOf(addrs[0].address);
-      let newOwnerBalanceAfter = await paymentToken.balanceOf(addrs[2].address);
-      let collectionOwnerBalAfter = await paymentToken.balanceOf(collectionOwnerAddress);
-
-      let devFeeAmount = ONE_ETH.mul(devFee).div(10000);
-      let beanieHolderFeeAmount = ONE_ETH.mul(beanieHolderFee).div(10000);
-      let beanBuybackFeeAmount = ONE_ETH.mul(beanBuybackFee).div(10000);
-      let collectionOwnerFeeAmount = ONE_ETH.mul(collectionOwnerFee).div(10000);
-      let afterFeePrice = ONE_ETH.mul(totalFee).div(10000);
-
-      //Check principal balances
-      expect(oldOwnerBalanceAfter).to.eql(oldOwnerBalanceBefore.add(ONE_ETH).sub(afterFeePrice));
-      expect(newOwnerBalanceAfter).to.eql(newOwnerBalanceBefore.sub(ONE_ETH));
-
-      expect(await paymentToken.balanceOf(devAddress)).to.eql(devFeeAmount);
-      expect(await paymentToken.balanceOf(beanHolderAddress)).to.eql(beanieHolderFeeAmount);
-      expect(await paymentToken.balanceOf(beanBuybackAddress)).to.eql(beanBuybackFeeAmount);
-      expect(await collectionOwnerBalAfter).to.eql(collectionOwnerBalBefore.add(collectionOwnerFeeAmount));
-
-      oldOwnerBalanceBefore = await paymentToken.balanceOf(addrs[0].address);
-      newOwnerBalanceBefore = await paymentToken.balanceOf(addrs[3].address);
-      collectionOwnerBalBefore = await paymentToken.balanceOf(collectionOwnerAddress);
-
-      await beanieMarket.connect(addrs[0]).acceptOffer(offer1Hash);
-
-      oldOwnerBalanceAfter = await paymentToken.balanceOf(addrs[0].address);
-      newOwnerBalanceAfter = await paymentToken.balanceOf(addrs[3].address);
-      collectionOwnerBalAfter = await paymentToken.balanceOf(collectionOwnerAddress);
-
-      //Check principal balances
-      expect(oldOwnerBalanceAfter).to.eql(oldOwnerBalanceBefore.add(ONE_ETH).sub(afterFeePrice));
-      expect(newOwnerBalanceAfter).to.eql(newOwnerBalanceBefore.sub(ONE_ETH));
-
-      expect(await paymentToken.balanceOf(devAddress)).to.eql(devFeeAmount.mul(2));
-      expect(await paymentToken.balanceOf(beanHolderAddress)).to.eql(beanieHolderFeeAmount.mul(2));
-      expect(await paymentToken.balanceOf(beanBuybackAddress)).to.eql(beanBuybackFeeAmount.mul(2));
-      expect(await collectionOwnerBalAfter).to.eql(collectionOwnerBalBefore.add(collectionOwnerFeeAmount));
-
-    });
-
-    it("Fulfill non-escrow feesOn sends correct token amount, autosend off and then process", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeOffersFixture);
-
-      await beanieMarket.connect(owner).setFeesOn(true);
-      await beanieMarket.connect(owner).setAutoSendFees(false);
-
-      const devFee = await beanieMarket.devFee();
-      const beanieHolderFee = await beanieMarket.beanieHolderFee();
-      const beanBuybackFee = await beanieMarket.beanBuybackFee();
-      const collectionOwnerFee = await beanieMarket.getCollectionFee(dummyNFT.address);
-      const totalFee = devFee.add(beanieHolderFee).add(beanBuybackFee).add(collectionOwnerFee);
-
-      const devAddress = await beanieMarket.devAddress();
-      const beanHolderAddress = await beanieMarket.beanieHolderAddress();
-      const beanBuybackAddress = await beanieMarket.beanBuybackAddress();
-      const collectionOwnerAddress = await beanieMarket.getCollectionOwner(dummyNFT.address);
-
-      let addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      let addr3offers = await beanieMarket.getOffersByOfferer(addrs[3].address);
-
-      const offer0Hash = addr2offers[0];
-      const offer0Data = await beanieMarket.offers(addr2offers[0]);
-
-      const offer1Hash = addr3offers[1];
-      const offer1Data = await beanieMarket.offers(addr3offers[1]);
-
-      await dummyNFT.connect(addrs[0]).setApprovalForAll(beanieMarket.address, true);
-
-      let oldOwnerBalanceBefore = await paymentToken.balanceOf(addrs[0].address);
-      let newOwnerBalanceBefore = await paymentToken.balanceOf(addrs[2].address);
-      let collectionOwnerBalBefore = await paymentToken.balanceOf(collectionOwnerAddress);
-
-      await beanieMarket.connect(addrs[0]).acceptOffer(offer0Hash);
-
-      let oldOwnerBalanceAfter = await paymentToken.balanceOf(addrs[0].address);
-      let newOwnerBalanceAfter = await paymentToken.balanceOf(addrs[2].address);
-      let collectionOwnerBalAfter = await paymentToken.balanceOf(collectionOwnerAddress);
-
-      let devFeeAmount = ONE_ETH.mul(devFee).div(10000);
-      let beanieHolderFeeAmount = ONE_ETH.mul(beanieHolderFee).div(10000);
-      let beanBuybackFeeAmount = ONE_ETH.mul(beanBuybackFee).div(10000);
-      let collectionOwnerFeeAmount = ONE_ETH.mul(collectionOwnerFee).div(10000);
-      let afterFeePrice = ONE_ETH.mul(totalFee).div(10000);
-
-      //Check principal balances
-      expect(oldOwnerBalanceAfter).to.eql(oldOwnerBalanceBefore.add(ONE_ETH).sub(afterFeePrice));
-      expect(newOwnerBalanceAfter).to.eql(newOwnerBalanceBefore.sub(ONE_ETH));
-
-      await beanieMarket.connect(owner).processDevFeesToken();
-
-      expect(await paymentToken.balanceOf(devAddress)).to.eql(devFeeAmount);
-      expect(await paymentToken.balanceOf(beanHolderAddress)).to.eql(beanieHolderFeeAmount);
-      expect(await paymentToken.balanceOf(beanBuybackAddress)).to.eql(beanBuybackFeeAmount);
-      expect(await collectionOwnerBalAfter).to.eql(collectionOwnerBalBefore.add(collectionOwnerFeeAmount));
-
-      oldOwnerBalanceBefore = await paymentToken.balanceOf(addrs[0].address);
-      newOwnerBalanceBefore = await paymentToken.balanceOf(addrs[3].address);
-      collectionOwnerBalBefore = await paymentToken.balanceOf(collectionOwnerAddress);
-
-      await beanieMarket.connect(addrs[0]).acceptOffer(offer1Hash);
-
-      oldOwnerBalanceAfter = await paymentToken.balanceOf(addrs[0].address);
-      newOwnerBalanceAfter = await paymentToken.balanceOf(addrs[3].address);
-      collectionOwnerBalAfter = await paymentToken.balanceOf(collectionOwnerAddress);
-
-      //Check principal balances
-      expect(oldOwnerBalanceAfter).to.eql(oldOwnerBalanceBefore.add(ONE_ETH).sub(afterFeePrice));
-      expect(newOwnerBalanceAfter).to.eql(newOwnerBalanceBefore.sub(ONE_ETH));
-
-      await beanieMarket.connect(owner).processDevFeesToken();
-
-      expect(await paymentToken.balanceOf(devAddress)).to.eql(devFeeAmount.mul(2));
-      expect(await paymentToken.balanceOf(beanHolderAddress)).to.eql(beanieHolderFeeAmount.mul(2));
-      expect(await paymentToken.balanceOf(beanBuybackAddress)).to.eql(beanBuybackFeeAmount.mul(2));
-      expect(await collectionOwnerBalAfter).to.eql(collectionOwnerBalBefore.add(collectionOwnerFeeAmount));
-
-    });
-  });
-
-  describe("Make non-escrow offer errors", function() {
-    it("Cannot make offer with zero price", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeEscrowOffersFixture);
-      await expect(beanieMarket.connect(addrs[0]).makeOffer(dummyNFT.address, 4, 0, now)
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANZeroPrice");
-    });
-
-    it("Cannot make offer if contract is not approved", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeEscrowOffersFixture);
-      await expect(beanieMarket.connect(addrs[0]).makeOffer(dummyNFT.address, 4, ONE_ETH, now)
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANContractNotApproved");
-    });
-
-    it("Cannot make offer if account does not own enough paymentToken", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeEscrowOffersFixture);
-
-      await paymentToken.connect(addrs[11]).approve(beanieMarket.address, ethers.constants.MaxUint256);
-
-      await expect(beanieMarket.connect(addrs[11]).makeOffer(dummyNFT.address, 4, ONE_ETH, now)
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANUserTokensLow");
-    });
-
-    it("Cannot make offer with expiry before now", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeEscrowOffersFixture);
-
-      await paymentToken.connect(addrs[0]).approve(beanieMarket.address, ethers.constants.MaxUint256);
-
-      await expect(beanieMarket.connect(addrs[0]).makeOffer(dummyNFT.address, 4, ONE_ETH, now-1)
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANBadExpiry");
-      await expect(beanieMarket.connect(addrs[0]).makeOffer(dummyNFT.address, 4, ONE_ETH, now)
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANBadExpiry");
-    });
-  });
-
-  describe("Fulfill non-escrow errors", function () {
-
-    it("Cannot accept non-existent offer", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeOffersFixture);
-      await expect(beanieMarket.connect(addrs[2]).acceptOffer(ethers.utils.hexZeroPad(0x2, 32))
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANCollectionNotEnabled");
-    });
-
-    it("Cannot accept offer past expiry", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeOffersFixture);
-
-      const addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-
-      const offer0Hash = addr2offers[0];
-      const offer0Data = await beanieMarket.offers(addr2offers[0]);
-
-      await network.provider.send("evm_setNextBlockTimestamp", [Number(offer0Data.expiry) + 10])
-      await network.provider.send("evm_mine")
-
-      await dummyNFT.connect(addrs[0]).setApprovalForAll(beanieMarket.address, true);
-      await expect(beanieMarket.connect(addrs[2]).acceptOffer(offer0Hash)
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANOrderExpired");
-    });
-
-    it("Cannot accept offer if not current token owner", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeOffersFixture);
-
-      const addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-
-      const offer0Hash = addr2offers[0];
-      const offer0Data = await beanieMarket.offers(addr2offers[0]);
-
-      await dummyNFT.connect(addrs[0]).setApprovalForAll(beanieMarket.address, true);
-      await expect(beanieMarket.connect(addrs[5]).acceptOffer(offer0Hash)
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANCallerNotOwner");
-    });
-
-  });
-
-  describe("Cancel non-escrow offer", function () {
-    it("Offerer can cancel non-escrow offer", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeOffersFixture);
-
-      let addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      const addr2offersTail = await addr2offers[addr2offers.length - 1];
-
-      let offerHash = addr2offers[0];
-
-      await beanieMarket.connect(addrs[2]).cancelOffer(offerHash);
-
-      addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-
-      expect(addr2offers).to.not.contain(offerHash)
-      expect(await beanieMarket.posInOffers(addr2offersTail)).to.eql(BIG_ZERO);
-      expect(await beanieMarket.offers(offerHash)).to.eql([BIG_ZERO, BIG_ZERO, BIG_ZERO, ADDR_ZERO, ADDR_ZERO, false]);
-    });
-
-    it("Token owner can cancel non-escrow offer", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeOffersFixture);
-
-      let addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      const addr2offersTail = await addr2offers[addr2offers.length - 1];
-
-      let offerHash = addr2offers[0];
-
-      await beanieMarket.connect(addrs[0]).cancelOffer(offerHash);
-
-      addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-
-      expect(addr2offers).to.not.contain(offerHash)
-      expect(await beanieMarket.posInOffers(addr2offersTail)).to.eql(BIG_ZERO);
-      expect(await beanieMarket.offers(offerHash)).to.eql([BIG_ZERO, BIG_ZERO, BIG_ZERO, ADDR_ZERO, ADDR_ZERO, false]);
-    });
-
-    it("Admin can cancel non-escrow offer", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeOffersFixture);
-
-      let addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      const addr2offersTail = await addr2offers[addr2offers.length - 1];
-
-      let offerHash = addr2offers[0];
-
-      await beanieMarket.connect(owner).cancelOffer(offerHash);
-
-      addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-
-      expect(addr2offers).to.not.contain(offerHash)
-      expect(await beanieMarket.posInOffers(addr2offersTail)).to.eql(BIG_ZERO);
-      expect(await beanieMarket.offers(offerHash)).to.eql([BIG_ZERO, BIG_ZERO, BIG_ZERO, ADDR_ZERO, ADDR_ZERO, false]);
-    });
-
-    it("Anyone can cancel non-escrow offer if past expiry", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeOffersFixture);
-
-      let addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      const addr2offersTail = await addr2offers[addr2offers.length - 1];
-
-      let offerHash = addr2offers[0];
-      const offer0Data = await beanieMarket.offers(offerHash);
-      const expiry = offer0Data.expiry
-
-      await expect(beanieMarket.connect(addrs[5]).cancelOffer(offerHash)
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANNotAuthorized");
-
-      await network.provider.send("evm_setNextBlockTimestamp", [Number(expiry) + 10])
-      await network.provider.send("evm_mine")
-
-      await beanieMarket.connect(addrs[8]).cancelOffer(offerHash);
-
-      addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-
-      expect(addr2offers).to.not.contain(offerHash)
-      expect(await beanieMarket.posInOffers(addr2offersTail)).to.eql(BIG_ZERO);
-      expect(await beanieMarket.offers(offerHash)).to.eql([BIG_ZERO, BIG_ZERO, BIG_ZERO, ADDR_ZERO, ADDR_ZERO, false]);
-    });
-  });
-
-  describe("Cancel non-escrow offer errors", function () {
-    it("Address cannot cancel if not offerer, token owner, admin, or past expiry", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeOffersFixture);
-
-      let addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      const addr2offersTail = await addr2offers[addr2offers.length - 1];
-
-      let offerHash = addr2offers[0];
-      const offer0Data = await beanieMarket.offers(offerHash);
-      const expiry = offer0Data.expiry
-
-      await expect(beanieMarket.connect(addrs[5]).cancelOffer(offerHash)
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANNotAuthorized");
-    });
-  });
-
-  describe("escrow offers", function () {
-    it("Make escrow offer", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndNFTFixture);
-
-      //Check the storage structures of several listings. We do this for two different lister addresses to test.
-      //TODO: maybe add a second contract address.
-
-      let balBefore = await addrs[2].getBalance();
-      let tx = await beanieMarket.connect(addrs[2]).makeEscrowedOffer(dummyNFT.address, 1, now + 10, {value: ONE_ETH}); // offer1
-      let balAfter = await addrs[2].getBalance();
-      let receipt = await tx.wait();
-      let gasSpent = receipt.gasUsed.mul(receipt.effectiveGasPrice);
-
-      expect(balAfter).to.eql(balBefore.sub(ONE_ETH).sub(gasSpent));
-
-      const addr2Offers_1 = await beanieMarket.getOffersByOfferer(addrs[2].address)
-      expect(addr2Offers_1.length).to.equal(1);
-
-      //same params, make sure it works
-      await beanieMarket.connect(addrs[2]).makeEscrowedOffer(dummyNFT.address, 1, now + 10, {value: ONE_ETH});
-      const addr2Offers_2 = await beanieMarket.getOffersByOfferer(addrs[2].address)
-      expect(addr2Offers_2.length).to.equal(2);
-      expect(addr2Offers_2[0]).to.not.equal(addr2Offers_2[1]);
-
-      //Offer for a different token
-      balBefore = await addrs[2].getBalance();
-      tx = await beanieMarket.connect(addrs[2]).makeEscrowedOffer(dummyNFT.address, 2, now + 100, {value: ONE_ETH.mul(2)}); //offer2
-      balAfter = await addrs[2].getBalance();
-      receipt = await tx.wait();
-      gasSpent = receipt.gasUsed.mul(receipt.effectiveGasPrice);
-
-      expect(balAfter).to.eql(balBefore.sub(ONE_ETH.mul(2)).sub(gasSpent));
-
-      const addr2Offers_3 = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      expect(addr2Offers_3.length).to.equal(3);
-
-      //3rd account makes an offer
-      await beanieMarket.connect(addrs[3]).makeEscrowedOffer(dummyNFT.address, 1, now + 1000, {value: ONE_ETH}); //offer3
-      const addr3Offers_1 = await beanieMarket.getOffersByOfferer(addrs[3].address);
-      expect(addr3Offers_1.length).to.equal(1);
-
-      const offer1 = await beanieMarket.offers(addr2Offers_3[0]);
-      const offer2 = await beanieMarket.offers(addr2Offers_3[2]);
-      const offer3 = await beanieMarket.offers(addr3Offers_1[0]);
-
-      expect(offer1).to.eql(
-        [
-          ethers.BigNumber.from(1),
-          ONE_ETH,
-          ethers.BigNumber.from(now + 10),
-          dummyNFT.address,
-          addrs[2].address,
-          true
-        ]
-      )
-
-      expect(offer2).to.eql(
-        [
-          ethers.BigNumber.from(2),
-          ONE_ETH.mul(2),
-          ethers.BigNumber.from(now + 100),
-          dummyNFT.address,
-          addrs[2].address,
-          true
-        ]
-      )
-
-      expect(offer3).to.eql(
-        [
-          ethers.BigNumber.from(1),
-          ONE_ETH,
-          ethers.BigNumber.from(now + 1000),
-          dummyNFT.address,
-          addrs[3].address,
-          true
-        ]
-      )
-    });
-
-    //TODO: Chain offer fulfillment
-    it("Fulfill escrow offer ownership change", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeEscrowOffersFixture);
-
-      const addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      const addr3offers = await beanieMarket.getOffersByOfferer(addrs[3].address);
-
-      const offer0Hash = addr2offers[0];
-      const offer0Data = await beanieMarket.offers(addr2offers[0]);
-      const offer0OldOwner = await dummyNFT.ownerOf(offer0Data.tokenId);
-
-      await dummyNFT.connect(addrs[0]).setApprovalForAll(beanieMarket.address, true);
-      await beanieMarket.connect(addrs[0]).acceptOffer(offer0Hash);
-      expect(await dummyNFT.ownerOf(offer0Data.tokenId)).to.equal(offer0Data.offerer);
-
-      const offer1Hash = addr3offers[1];
-      const offer1Data = await beanieMarket.offers(offer1Hash);
-      const offer1OldOwner = await dummyNFT.ownerOf(offer1Data.tokenId);
-
-      await beanieMarket.connect(addrs[0]).acceptOffer(offer1Hash);
-      expect(await dummyNFT.ownerOf(offer1Data.tokenId)).to.equal(offer1Data.offerer);
-    });
-
-    it("Fulfill escrow offer update storage structures", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeEscrowOffersFixture);
-
-      let addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      let addr3offers = await beanieMarket.getOffersByOfferer(addrs[3].address);
-
-      const addr2offersTail = await addr2offers[addr2offers.length - 1];
-      const addr3offersTail = await addr3offers[addr3offers.length - 1];
-
-      const offer0Hash = addr2offers[0];
-      const offer0Data = await beanieMarket.offers(offer0Hash);
-
-      const offer1Hash = addr3offers[1];
-      const offer1Data = await beanieMarket.offers(offer1Hash);
-
-      await dummyNFT.connect(addrs[0]).setApprovalForAll(beanieMarket.address, true);
-      await beanieMarket.connect(addrs[0]).acceptOffer(offer0Hash);
-      await beanieMarket.connect(addrs[0]).acceptOffer(offer1Hash);
-
-      addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      addr3offers = await beanieMarket.getOffersByOfferer(addrs[3].address);
-      expect(addr2offers).to.not.contain(offer0Hash)
-      expect(addr3offers).to.not.contain(offer1Hash)
-
-      expect(await beanieMarket.posInOffers(addr2offersTail)).to.eql(BIG_ZERO);
-
-    });
-
-    //TODO: reinspect these
-    it("Fulfill escrow offer update storage structures (single length array)", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeEscrowOffersFixture);
-
-      let addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      let addr3offers = await beanieMarket.getOffersByOfferer(addrs[3].address);
-
-      const addr2offersTail = await addr2offers[addr2offers.length - 1];
-      const addr3offersTail = await addr3offers[addr3offers.length - 1];
-
-      const offer0Hash = addr2offers[0];
-      const offer0Data = await beanieMarket.offers(addr2offers[0]);
-
-      const offer1Hash = addr3offers[1];
-      const offer1Data = await beanieMarket.offers(addr3offers[0]);
-
-      await dummyNFT.connect(addrs[0]).setApprovalForAll(beanieMarket.address, true);
-      await beanieMarket.connect(addrs[0]).acceptOffer(offer0Hash);
-      await beanieMarket.connect(addrs[0]).acceptOffer(offer1Hash);
-
-      addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      addr3offers = await beanieMarket.getOffersByOfferer(addrs[3].address);
-      expect(addr2offers).to.not.contain(offer0Hash)
-      expect(addr3offers).to.not.contain(offer1Hash)
-
-      expect(await beanieMarket.posInOffers(addr2offersTail)).to.eql(BIG_ZERO);
-
-    });
-
-    it("Fulfill escrow offer feesOff sends correct eth amount", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeEscrowOffersFixture);
-
-      await beanieMarket.connect(owner).setFeesOn(false);
-
-      let addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      let addr3offers = await beanieMarket.getOffersByOfferer(addrs[3].address);
-
-      const offer0Hash = addr2offers[0];
-      const offer0Data = await beanieMarket.offers(addr2offers[0]);
-
-      const offer1Hash = addr3offers[1];
-      const offer1Data = await beanieMarket.offers(addr3offers[1]);
-
-      await dummyNFT.connect(addrs[0]).setApprovalForAll(beanieMarket.address, true);
-
-      let oldOwnerBalanceBefore = await addrs[0].getBalance();
-      let newOwnerBalanceBefore = await addrs[2].getBalance();
-
-      let tx = await beanieMarket.connect(addrs[0]).acceptOffer(offer0Hash);
-      let receipt = await tx.wait();
-      let gasSpent = receipt.gasUsed.mul(receipt.effectiveGasPrice);
-
-      let oldOwnerBalanceAfter = await addrs[0].getBalance();
-      let newOwnerBalanceAfter = await addrs[2].getBalance();
-
-      expect(oldOwnerBalanceAfter).to.eql(oldOwnerBalanceBefore.add(ONE_ETH).sub(gasSpent));
-      expect(newOwnerBalanceAfter).to.eql(newOwnerBalanceBefore);
-
-      oldOwnerBalanceBefore = await addrs[0].getBalance();
-
-      tx = await beanieMarket.connect(addrs[0]).acceptOffer(offer1Hash);
-      receipt = await tx.wait();
-      gasSpent = receipt.gasUsed.mul(receipt.effectiveGasPrice);
-
-      oldOwnerBalanceAfter = await addrs[0].getBalance();
-
-      expect(oldOwnerBalanceAfter).to.eql(oldOwnerBalanceBefore.add(ONE_ETH).sub(gasSpent));
-    });
-
-    it("Fulfill escrow feesOn sends correct token amount, autosend on", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeEscrowOffersFixture);
-
-      await beanieMarket.connect(owner).setFeesOn(true);
-      await beanieMarket.connect(owner).setAutoSendFees(true);
-
-      const devFee = await beanieMarket.devFee();
-      const beanieHolderFee = await beanieMarket.beanieHolderFee();
-      const beanBuybackFee = await beanieMarket.beanBuybackFee();
-      const collectionOwnerFee = await beanieMarket.getCollectionFee(dummyNFT.address);
-      const totalFee = devFee.add(beanieHolderFee).add(beanBuybackFee).add(collectionOwnerFee);
-
-      const devAddress = await beanieMarket.devAddress();
-      const beanHolderAddress = await beanieMarket.beanieHolderAddress();
-      const beanBuybackAddress = await beanieMarket.beanBuybackAddress();
-      const collectionOwnerAddress = await beanieMarket.getCollectionOwner(dummyNFT.address);
-
-      let addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      let addr3offers = await beanieMarket.getOffersByOfferer(addrs[3].address);
-
-      const offer0Hash = addr2offers[0];
-      const offer0Data = await beanieMarket.offers(addr2offers[0]);
-      const offer0price = offer0Data.price;
-
-      const offer1Hash = addr3offers[1];
-      const offer1Data = await beanieMarket.offers(addr3offers[1]);
-      const offer1price = offer1Data.price;
-
-      await dummyNFT.connect(addrs[0]).setApprovalForAll(beanieMarket.address, true);
-      let oldOwnerBalanceBefore = await addrs[0].getBalance();
-      let collectionOwnerBalBefore = await ethers.provider.getBalance(collectionOwnerAddress);
-
-      let tx = await beanieMarket.connect(addrs[0]).acceptOffer(offer0Hash);
-      let receipt = await tx.wait();
-      let gasSpent = receipt.gasUsed.mul(receipt.effectiveGasPrice);
-
-      let oldOwnerBalanceAfter = await addrs[0].getBalance();
-      let collectionOwnerBalAfter = await ethers.provider.getBalance(collectionOwnerAddress);
-
-      let devFeeAmount = offer0price.mul(devFee).div(10000);
-      let beanieHolderFeeAmount = offer0price.mul(beanieHolderFee).div(10000);
-      let beanBuybackFeeAmount = offer0price.mul(beanBuybackFee).div(10000);
-      let collectionOwnerFeeAmount = offer0price.mul(collectionOwnerFee).div(10000);
-      let afterFeePrice = offer0price.mul(totalFee).div(10000);
-
-      //Check principal balances
-      expect(oldOwnerBalanceAfter).to.eql(oldOwnerBalanceBefore.add(offer0price).sub(afterFeePrice).sub(gasSpent));
-
-      expect(await ethers.provider.getBalance(devAddress)).to.eql(devFeeAmount);
-      expect(await ethers.provider.getBalance(beanHolderAddress)).to.eql(beanieHolderFeeAmount);
-      expect(await ethers.provider.getBalance(beanBuybackAddress)).to.eql(beanBuybackFeeAmount);
-      expect(await collectionOwnerBalAfter).to.eql(collectionOwnerBalBefore.add(collectionOwnerFeeAmount));
-
-      oldOwnerBalanceBefore = await addrs[0].getBalance();
-      collectionOwnerBalBefore = await ethers.provider.getBalance(collectionOwnerAddress);
-
-      tx = await beanieMarket.connect(addrs[0]).acceptOffer(offer1Hash);
-      receipt = await tx.wait();
-      gasSpent = receipt.gasUsed.mul(receipt.effectiveGasPrice);
-
-      let devFeeAmount_1 = offer1price.mul(devFee).div(10000);
-      let beanieHolderFeeAmount_1 = offer1price.mul(beanieHolderFee).div(10000);
-      let beanBuybackFeeAmount_1 = offer1price.mul(beanBuybackFee).div(10000);
-      let collectionOwnerFeeAmount_1 = offer1price.mul(collectionOwnerFee).div(10000);
-      let afterFeePrice_1 = offer1price.mul(totalFee).div(10000);
-
-      oldOwnerBalanceAfter = await addrs[0].getBalance();
-      collectionOwnerBalAfter = await ethers.provider.getBalance(collectionOwnerAddress);
-      
-      //Check principal balances
-      expect(oldOwnerBalanceAfter).to.eql(oldOwnerBalanceBefore.add(offer0price).sub(afterFeePrice_1).sub(gasSpent));
-
-      expect(await ethers.provider.getBalance(devAddress)).to.eql(devFeeAmount.add(devFeeAmount_1));
-      expect(await ethers.provider.getBalance(beanHolderAddress)).to.eql(beanieHolderFeeAmount.add(beanieHolderFeeAmount_1));
-      expect(await ethers.provider.getBalance(beanBuybackAddress)).to.eql(beanBuybackFeeAmount.add(beanBuybackFeeAmount_1));
-      expect(await collectionOwnerBalAfter).to.eql(collectionOwnerBalBefore.add(collectionOwnerFeeAmount_1));
-    });
-
-    it("Fulfill escrow feesOn sends correct token amount, autosend off and then process", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeEscrowOffersFixture);
-
-      await beanieMarket.connect(owner).setFeesOn(true);
-      await beanieMarket.connect(owner).setAutoSendFees(false);
-
-      const devFee = await beanieMarket.devFee();
-      const beanieHolderFee = await beanieMarket.beanieHolderFee();
-      const beanBuybackFee = await beanieMarket.beanBuybackFee();
-      const collectionOwnerFee = await beanieMarket.getCollectionFee(dummyNFT.address);
-      const totalFee = devFee.add(beanieHolderFee).add(beanBuybackFee).add(collectionOwnerFee);
-
-      const devAddress = await beanieMarket.devAddress();
-      const beanHolderAddress = await beanieMarket.beanieHolderAddress();
-      const beanBuybackAddress = await beanieMarket.beanBuybackAddress();
-      const collectionOwnerAddress = await beanieMarket.getCollectionOwner(dummyNFT.address);
-
-      let addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      let addr3offers = await beanieMarket.getOffersByOfferer(addrs[3].address);
-
-      const offer0Hash = addr2offers[0];
-      const offer0Data = await beanieMarket.offers(addr2offers[0]);
-      const offer0price = offer0Data.price;
-
-      const offer1Hash = addr3offers[1];
-      const offer1Data = await beanieMarket.offers(addr3offers[1]);
-      const offer1price = offer1Data.price;
-
-      await dummyNFT.connect(addrs[0]).setApprovalForAll(beanieMarket.address, true);
-      let oldOwnerBalanceBefore = await addrs[0].getBalance();
-      let collectionOwnerBalBefore = await ethers.provider.getBalance(collectionOwnerAddress);
-
-      let tx = await beanieMarket.connect(addrs[0]).acceptOffer(offer0Hash);
-      let receipt = await tx.wait();
-      let gasSpent = receipt.gasUsed.mul(receipt.effectiveGasPrice);
-
-      let oldOwnerBalanceAfter = await addrs[0].getBalance();
-      let collectionOwnerBalAfter = await ethers.provider.getBalance(collectionOwnerAddress);
-
-      let devFeeAmount = offer0price.mul(devFee).div(10000);
-      let beanieHolderFeeAmount = offer0price.mul(beanieHolderFee).div(10000);
-      let beanBuybackFeeAmount = offer0price.mul(beanBuybackFee).div(10000);
-      let collectionOwnerFeeAmount = offer0price.mul(collectionOwnerFee).div(10000);
-      let afterFeePrice = offer0price.mul(totalFee).div(10000);
-
-      //Check principal balances
-      expect(oldOwnerBalanceAfter).to.eql(oldOwnerBalanceBefore.add(offer0price).sub(afterFeePrice).sub(gasSpent));
-      expect(await collectionOwnerBalAfter).to.eql(collectionOwnerBalBefore.add(collectionOwnerFeeAmount));
-      expect(await ethers.provider.getBalance(devAddress)).to.eql(BIG_ZERO);
-      expect(await ethers.provider.getBalance(beanHolderAddress)).to.eql(BIG_ZERO);
-      expect(await ethers.provider.getBalance(beanBuybackAddress)).to.eql(BIG_ZERO);
-
-      await beanieMarket.connect(owner).processDevFeesEth();
-
-      expect(await ethers.provider.getBalance(devAddress)).to.eql(devFeeAmount);
-      expect(await ethers.provider.getBalance(beanHolderAddress)).to.eql(beanieHolderFeeAmount);
-      expect(await ethers.provider.getBalance(beanBuybackAddress)).to.eql(beanBuybackFeeAmount);
-
-      oldOwnerBalanceBefore = await addrs[0].getBalance();
-      collectionOwnerBalBefore = await ethers.provider.getBalance(collectionOwnerAddress);
-
-      tx = await beanieMarket.connect(addrs[0]).acceptOffer(offer1Hash);
-      receipt = await tx.wait();
-      gasSpent = receipt.gasUsed.mul(receipt.effectiveGasPrice);
-
-      let devFeeAmount_1 = offer1price.mul(devFee).div(10000);
-      let beanieHolderFeeAmount_1 = offer1price.mul(beanieHolderFee).div(10000);
-      let beanBuybackFeeAmount_1 = offer1price.mul(beanBuybackFee).div(10000);
-      let collectionOwnerFeeAmount_1 = offer1price.mul(collectionOwnerFee).div(10000);
-      let afterFeePrice_1 = offer1price.mul(totalFee).div(10000);
-
-      oldOwnerBalanceAfter = await addrs[0].getBalance();
-      collectionOwnerBalAfter = await ethers.provider.getBalance(collectionOwnerAddress);
-      
-      //Check principal balances
-      expect(oldOwnerBalanceAfter).to.eql(oldOwnerBalanceBefore.add(offer0price).sub(afterFeePrice_1).sub(gasSpent));
-      expect(await collectionOwnerBalAfter).to.eql(collectionOwnerBalBefore.add(collectionOwnerFeeAmount_1));
-
-      await beanieMarket.connect(owner).processDevFeesEth();
-      
-      expect(await ethers.provider.getBalance(devAddress)).to.eql(devFeeAmount.add(devFeeAmount_1));
-      expect(await ethers.provider.getBalance(beanHolderAddress)).to.eql(beanieHolderFeeAmount.add(beanieHolderFeeAmount_1));
-      expect(await ethers.provider.getBalance(beanBuybackAddress)).to.eql(beanBuybackFeeAmount.add(beanBuybackFeeAmount_1));
-    });
-  });
-
-  describe("Make escrow offer errors", function() {
-    it("Cannot make offer with zero price", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeEscrowOffersFixture);
-      await expect(beanieMarket.connect(addrs[0]).makeEscrowedOffer(dummyNFT.address, 4, now, {value: ethers.constants.Zero})
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANZeroPrice");
-    });
-
-    it("Cannot make offer with expiry before now", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeEscrowOffersFixture);
-
-      await expect(beanieMarket.connect(addrs[0]).makeEscrowedOffer(dummyNFT.address, 4, now-1, {value: ONE_ETH})
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANBadExpiry");
-      await expect(beanieMarket.connect(addrs[0]).makeEscrowedOffer(dummyNFT.address, 4, now, {value: ONE_ETH})
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANBadExpiry");
-    });
-  });
-
-  describe("Fulfill escrow offer errors", function () {
-    it("Cannot accept non-existent offer", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeEscrowOffersFixture);
-      await expect(beanieMarket.connect(addrs[2]).acceptOffer(ethers.utils.hexZeroPad(0x2, 32))
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANCollectionNotEnabled");
-    });
-
-    it("Cannot accept offer past expiry", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeEscrowOffersFixture);
-
-      const addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-
-      const offer0Hash = addr2offers[0];
-      const offer0Data = await beanieMarket.offers(addr2offers[0]);
-
-      await network.provider.send("evm_setNextBlockTimestamp", [Number(offer0Data.expiry) + 10])
-      await network.provider.send("evm_mine")
-
-      await dummyNFT.connect(addrs[0]).setApprovalForAll(beanieMarket.address, true);
-      await expect(beanieMarket.connect(addrs[2]).acceptOffer(offer0Hash)
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANOrderExpired");
-    });
-
-    it("Cannot accept offer if not current token owner", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeEscrowOffersFixture);
-
-      const addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-
-      const offer0Hash = addr2offers[0];
-      const offer0Data = await beanieMarket.offers(addr2offers[0]);
-
-      await dummyNFT.connect(addrs[0]).setApprovalForAll(beanieMarket.address, true);
-      await expect(beanieMarket.connect(addrs[5]).acceptOffer(offer0Hash)
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANCallerNotOwner");
-    });
-
-  });
-
-  describe("Cancel escrow offer", function () {
-    it("Offerer can cancel escrow offer, offerer recieves escrow back", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeEscrowOffersFixture);
-
-      let addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      const addr2offersTail = await addr2offers[addr2offers.length - 1];
-
-      let offerHash = addr2offers[0];
-      const offer0Data = await beanieMarket.offers(addr2offers[0]);
-      const offer0price = offer0Data.price;
-
-      let offererBalanceBefore = await addrs[2].getBalance();
-
-      let tx = await beanieMarket.connect(addrs[2]).cancelOffer(offerHash);
-      let receipt = await tx.wait();
-      let gasSpent = receipt.gasUsed.mul(receipt.effectiveGasPrice);
-
-      let offererBalanceAfter = await addrs[2].getBalance();
-
-      expect(offererBalanceAfter).to.eql(offererBalanceBefore.add(offer0price).sub(gasSpent));
-
-      addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      expect(addr2offers).to.not.contain(offerHash)
-      expect(await beanieMarket.posInOffers(addr2offersTail)).to.eql(BIG_ZERO);
-      expect(await beanieMarket.offers(offerHash)).to.eql([BIG_ZERO, BIG_ZERO, BIG_ZERO, ADDR_ZERO, ADDR_ZERO, false]);
-    });
-
-    it("Offerer can cancel escrow offer", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeEscrowOffersFixture);
-
-      let addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      const addr2offersTail = await addr2offers[addr2offers.length - 1];
-
-      let offerHash = addr2offers[0];
-
-      await expect(beanieMarket.connect(addrs[2]).cancelOffer(offerHash)).to.changeEtherBalances(
-        [beanieMarket.address, addrs[2].address], [ONE_ETH.mul(-1), ONE_ETH]);
-
-      addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-
-      expect(addr2offers).to.not.contain(offerHash)
-      expect(await beanieMarket.posInOffers(addr2offersTail)).to.eql(BIG_ZERO);
-      expect(await beanieMarket.offers(offerHash)).to.eql([BIG_ZERO, BIG_ZERO, BIG_ZERO, ADDR_ZERO, ADDR_ZERO, false]);
-    });
-
-    it("Token owner can cancel escrow offer", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeEscrowOffersFixture);
-
-      let addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      const addr2offersTail = await addr2offers[addr2offers.length - 1];
-
-      let offerHash = addr2offers[0];
-
-      await expect(beanieMarket.connect(addrs[0]).cancelOffer(offerHash)).to.changeEtherBalances(
-        [beanieMarket.address, addrs[2].address], [ONE_ETH.mul(-1), ONE_ETH]);;
-
-      addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-
-      expect(addr2offers).to.not.contain(offerHash)
-      expect(await beanieMarket.posInOffers(addr2offersTail)).to.eql(BIG_ZERO);
-      expect(await beanieMarket.offers(offerHash)).to.eql([BIG_ZERO, BIG_ZERO, BIG_ZERO, ADDR_ZERO, ADDR_ZERO, false]);
-    });
-
-    it("Admin can cancel escrow offer", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeEscrowOffersFixture);
-
-      let addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      const addr2offersTail = await addr2offers[addr2offers.length - 1];
-
-      let offerHash = addr2offers[0];
-
-      await expect(beanieMarket.connect(owner).cancelOffer(offerHash)).to.changeEtherBalances(
-        [beanieMarket.address, addrs[2].address], [ONE_ETH.mul(-1), ONE_ETH]);
-
-      addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-
-      expect(addr2offers).to.not.contain(offerHash)
-      expect(await beanieMarket.posInOffers(addr2offersTail)).to.eql(BIG_ZERO);
-      expect(await beanieMarket.offers(offerHash)).to.eql([BIG_ZERO, BIG_ZERO, BIG_ZERO, ADDR_ZERO, ADDR_ZERO, false]);
-    });
-
-    it("Anyone can cancel escrow offer if past expiry", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeEscrowOffersFixture);
-
-      let addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      const addr2offersTail = await addr2offers[addr2offers.length - 1];
-
-      let offerHash = addr2offers[0];
-      const offer0Data = await beanieMarket.offers(offerHash);
-      const expiry = offer0Data.expiry
-
-      await expect(beanieMarket.connect(addrs[5]).cancelOffer(offerHash)
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANNotAuthorized");
-
-      await network.provider.send("evm_setNextBlockTimestamp", [Number(expiry) + 10])
-      await network.provider.send("evm_mine")
-
-      await expect(beanieMarket.connect(addrs[8]).cancelOffer(offerHash)).to.changeEtherBalances(
-        [beanieMarket.address, addrs[2].address], [ONE_ETH.mul(-1), ONE_ETH]);
-
-      addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-
-      expect(addr2offers).to.not.contain(offerHash)
-      expect(await beanieMarket.posInOffers(addr2offersTail)).to.eql(BIG_ZERO);
-      expect(await beanieMarket.offers(offerHash)).to.eql([BIG_ZERO, BIG_ZERO, BIG_ZERO, ADDR_ZERO, ADDR_ZERO, false]);
-    });
-  });
-
-  describe("Cancel escrow offer errors", function () {
-    it("Address cannot cancel if not offerer, token owner, admin, or past expiry", async function () {
-      const { beanieMarket, dummyNFT, paymentToken, owner, addrs, now } = await loadFixture(deployMarketAndMakeEscrowOffersFixture);
-
-      let addr2offers = await beanieMarket.getOffersByOfferer(addrs[2].address);
-      const addr2offersTail = await addr2offers[addr2offers.length - 1];
-
-      let offerHash = addr2offers[0];
-      const offer0Data = await beanieMarket.offers(offerHash);
-      const expiry = offer0Data.expiry
-
-      await expect(beanieMarket.connect(addrs[5]).cancelOffer(offerHash)
-        ).to.be.revertedWithCustomError(beanieMarket, "BEANNotAuthorized");
-    });
-  });
-});
